Versão usando um vetor de unsigned char como parametro das funções

+++ trunk/C.03/src/Headers/ZoomCanvas.h
@@ -5,7 +5,7 @@ GtkWidget* zoom_canvas_new();
 int color, gray;
 double zoom;
 double current_zoom;
-int brilho, contraste;
+double brilho, contraste;
 gboolean zoom_canvas_on_expose_event(GtkWidget *widget, GdkEventExpose *event, gpointer data);
 void change_config();
 void *captura_clique();
